reset_softHSM: add support of Homebrew on arm64
The path to store tokens is different on arm64.

+++ test/reset_softHSM.sh
@@ -4,7 +4,9 @@ set -e
 
 # (re)create a PKCS#11 token using SoftHSM v2
 
-rm -rf /var/lib/softhsm/tokens/*
-rm -rf /usr/local/var/lib/softhsm/tokens/*
+[ -d  /var/lib/softhsm/tokens/ ] && rm -rf /var/lib/softhsm/tokens/*
+[ -d /usr/local/var/lib/softhsm/tokens ] && rm -rf /usr/local/var/lib/softhsm/tokens/*
+[ -d /opt/homebrew/var/lib/softhsm/tokens ] && rm -rf /opt/homebrew/var/lib/softhsm/tokens/*
 
 softhsm2-util --init-token --label "A token" --pin 1234 --so-pin 123456 --slot 0
+softhsm2-util --show-slots
